% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdkman4r-package.R
\docType{package}
\name{sdkman4r-package}
\alias{sdkman4r}
\alias{sdkman4r-package}
\title{sdkman4r: R Wrapper of The Software Development Kit Manager}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
